Move ack set util postion
